refactors tests a little
